prepare for more chunking methods
